pyproject.toml
#------------------------------------------------------------------------------#
# SEE: https://python-poetry.org/docs/pyproject/#poetry-and-pep-517
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"


#------------------------------------------------------------------------------#
[tool.poetry]
name = "petervaro"
version = "0.0.0"
description = "Personal Stuff"
authors = ["Peter Varo <hello@petervaro.com>"]
license = "GPL-3.0-or-later"
packages = [{include = "scripts"}]


#------------------------------------------------------------------------------#
[tool.poetry.dependencies]
python = "^3.9"
tomlkit = "~0.7"
pillow = "~8.0"
weasyprint = "51"
htmlmin = "~0.1"
dominate = "~2.6"


#------------------------------------------------------------------------------#
[tool.poetry.dev-dependencies]
pdbpp = "~0.10"
flask = "~1.1"


#------------------------------------------------------------------------------#
[tool.poetry.scripts]
build = "scripts.build:main"
server = "scripts.server:main"
